+++ packages/docs/react-docs/src/routes/components+/link+/demos/link-colors-demo.tsx
@@ -6,8 +6,12 @@ export function LinkColorsDemo(): ReactElement {
   return (
     <div className="flex flex-col gap-4">
       {/* preview */}
-      <Link>Default</Link>
-      <Link emphasis="neutral">Neutral</Link>
+      <Link>default</Link>
+      <Link emphasis="neutral">neutral</Link>
+      <Link emphasis="brand">brand</Link>
+      <div className="bg-persistent-black p-4">
+        <Link emphasis="white-persistent">white-persistent</Link>
+      </div>
       {/* preview */}
     </div>
   )